+++ R/tq_mutate.R
@@ -145,7 +145,15 @@ merge_two_tibbles <- function(tib1, tib2, mutate_fun) {
     # Merge results
     if (identical(nrow(tib1), nrow(tib2))) {
 
-        ret <- dplyr::bind_cols(tib1, tib2[,-1])
+        tib2 <- tib2[,-1] # Drop date column
+
+        # Replace bad names
+        tib2 <- replace_bad_names(tib2, mutate_fun)
+
+        tib2 <- replace_duplicate_colnames(tib1, tib2)
+
+        ret <- dplyr::bind_cols(tib1, tib2)
+
 
     } else {
 
@@ -157,3 +165,74 @@ merge_two_tibbles <- function(tib1, tib2, mutate_fun) {
 
 }
 
+
+replace_duplicate_colnames <- function(tib1, tib2) {
+
+    # Collect column names
+    name_list_tib1 <- colnames(tib1)
+    name_list_tib2 <- colnames(tib2)
+    name_list <- c(name_list_tib1, name_list_tib2)
+
+    duplicates_exist <- detect_duplicates(name_list)
+
+    # Iteratively add .1, .2, .3 ... onto end of column names
+    if (duplicates_exist) {
+
+        i <- 1
+
+        while (duplicates_exist) {
+
+            dup_names_stripped <-
+                stringr::str_split(name_list[duplicated(name_list)],
+                                   pattern = "\\.",
+                                   simplify = FALSE) %>%
+                sapply(function(x) x[[1]])
+
+            name_list[duplicated(name_list)] <-
+                stringr::str_c(dup_names_stripped, ".", i)
+
+            i <- i + 1
+
+            duplicates_exist <- detect_duplicates(name_list)
+
+        }
+
+        name_list_tib2 <- name_list[(ncol(tib1) + 1):length(name_list)]
+
+        colnames(tib2) <- name_list_tib2
+
+    }
+
+    tib2
+
+}
+
+detect_duplicates <- function(name_list) {
+
+    name_list %>%
+        duplicated() %>%
+        any()
+
+}
+
+# bad / restricted names are names that get selected unintetionally by OHLC functions
+replace_bad_names <- function(tib, fun_name) {
+
+    bad_names_regex <- "open|high|low|close|volume|adjusted|price"
+
+    name_list_tib <- stringr::str_to_lower(colnames(tib))
+
+    detect_bad_names <- stringr::str_detect(name_list_tib, bad_names_regex)
+
+    if (any(detect_bad_names)) {
+
+        len <- length(name_list_tib[detect_bad_names])
+        name_list_tib[detect_bad_names] <- rep(fun_name, length.out = len)
+
+    }
+
+    colnames(tib) <- name_list_tib
+
+    tib
+
+}